Param vals in EstCC from val to var for testing purposes
Because of the way the App trait works, the members declared in EstCC
are null when tests are run (see
http://stackoverflow.com/questions/15346600/field-inside-object-which-extends-app-trait-is-set-to-null-why-is-that-so)

To get around this for testing code in EstimateMI and EstimateCC (which
use the globally declared parameters inside EstCC), the parameter values
have to be explicitly assigned inside EstCC before the test is run. As a
result, the fields that are used by other code (including the parameter
fields and the rEngine field) must be set to var instead of val, so that
they can be assigned when the test is run.